pycocci: fix error when python is python3
When running under an Arch Linux system where python is python3, the
following error was observed:

  Traceback (most recent call last):
    File "/home/dfrey/projects/backports-linux/devel/pycocci", line 172, in <module>
      ret = _main()
    File "/home/dfrey/projects/backports-linux/devel/pycocci", line 158, in _main
      if args.jobs > 0:
  TypeError: '>' not supported between instances of 'NoneType' and 'int'

The problem is that args.jobs is None in the case that the -j or --jobs
argument wasn't passed to the command. In python3, an exception is
generated when performing a comparison between different types.

Signed-off-by: David Frey <dpfrey@gmail.com>
(cherry picked from commit 5ab7098a8abed0a0e677892bc51f0bba684a4ba4)